refactor: move ModelRegistryClient to lib/model_manager/registry/client.py
Per architect review: registry is a model_manager-level abstraction (alongside
interface/, schema/, packager/) not a pusher-specific concern. Other workflows
(serving, validation) can reuse it without depending on pusher.

- lib/model_manager/registry/client.py — canonical ModelRegistryClient ABC +
  RegisteredModel (renamed from registry_client.py to avoid path redundancy)
- lib/model_manager/registry/__init__.py — exports both types
- lib/model_manager/registry/tests/client_test.py — 10 tests
- lib/model_manager/__init__.py — top-level docstring for model_manager public API
- pusher/registry_client.py — thin re-export shim → model_manager/registry/client
- pusher/tests/test_registry_client.py — import updated to canonical path
