+++ python/michelangelo/lib/model_manager/registry/__init__.py
@@ -1,0 +1,7 @@
+"""Model registry client interface for the model manager."""
+
+# flake8: noqa:F401
+from michelangelo.lib.model_manager.registry.client import (
+    ModelRegistryClient,
+    RegisteredModel,
+)
